+++ kernel/server/controller/dynamic/page.go
@@ -6,44 +6,13 @@ import (
 
 	"publish/client"
 	"publish/config"
-	"publish/server/controller"
 	"publish/server/status"
 
 	"github.com/gin-gonic/gin"
 )
 
-type page struct {
-	controller.Page
-}
-
-var (
-	P *page
-)
-
-func init() {
-	P = &page{}
-}
-
-/* 解析 URL 参数 id */
-func (*page) ID(c *gin.Context) {
-	id := c.GetString("id")
-	root_id := c.GetString("root_id")
-
-	c.Request.URL.Path = "/block/" + root_id
-	if id == root_id { // 查询文档块
-		query := c.Request.URL.Query()
-		query.Del("id")
-		// fmt.Printf("%+v\n", query)
-		// fmt.Printf("%+v\n", query.Encode())
-		c.Request.URL.RawQuery = query.Encode()
-		// router.HandleContext(c)
-	}
-
-	c.Redirect(http.StatusMovedPermanently, c.Request.URL.String())
-}
-
 /* 解析路由参数 id */
-func (*page) Block(c *gin.Context) {
+func Block(c *gin.Context) {
 	var (
 		root_id string
 		r       *client.ResponseBody
@@ -52,7 +21,7 @@ func (*page) Block(c *gin.Context) {
 	)
 
 	/* 获得指定文档的 DOM */
-	root_id = c.GetString("id")
+	root_id = c.GetString("root_id")
 	r, err = client.GetBlockDomByID(client.C.R(), root_id, 0)
 	r, err_msg = client.Response(r, err)
 	if r == nil {