+++ kernel/server/controller/interface.go
@@ -1,8 +1,0 @@
-package controller
-
-import "github.com/gin-gonic/gin"
-
-type Page interface {
-	ID(c *gin.Context)
-	Block(c *gin.Context)
-}