+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 - 输出日志文件 | Output log file.
 - 使用 MVC 框架重构项目结构 | Use MVC framework to rebuild project structure.
 - 实现文档页面加载的动态缓存模式 | Implement document page loading dynamic cache mode.
+- 实现文档页面加载的静态加载模式 | Implement document page loading static mode.